Remove listUnresolvedReferences from api (core) and test it in workspace
