+++ packages/cli/test/commands/element.test.ts
@@ -14,9 +14,8 @@
 * limitations under the License.
 */
 import open from 'open'
-import * as core from '@salto-io/core'
 import { ElemID, ObjectType, CORE_ANNOTATIONS } from '@salto-io/adapter-api'
-import { errors } from '@salto-io/workspace'
+import { errors, UnresolvedElemIDs } from '@salto-io/workspace'
 import { CliExitCode } from '../../src/types'
 import { cloneAction, moveToEnvsAction, moveToCommonAction, listUnresolvedAction, openAction } from '../../src/commands/element'
 import * as mocks from '../mocks'
@@ -25,7 +24,7 @@ import Prompts from '../../src/prompts'
 import { formatTargetEnvRequired } from '../../src/formatter'
 
 
-const mockedList: typeof core.listUnresolvedReferences = (_workspace, completeFromEnv) => (
+const mockedList = async (completeFromEnv?: string): Promise<UnresolvedElemIDs> => (
   completeFromEnv !== undefined
     ? Promise.resolve({
       found: [new ElemID('salesforce', 'aaa'), new ElemID('salesforce', 'bbb', 'instance', 'ccc')],
@@ -37,10 +36,6 @@ const mockedList: typeof core.listUnresolvedReferences = (_workspace, completeFr
     })
 )
 jest.mock('open')
-jest.mock('@salto-io/core', () => ({
-  ...jest.requireActual<{}>('@salto-io/core'),
-  listUnresolvedReferences: jest.fn().mockImplementation((_ws, env) => mockedList(_ws, env)),
-}))
 describe('Element command group', () => {
   describe('Clone command', () => {
     const cloneName = 'clone'
@@ -73,6 +68,7 @@ describe('Element command group', () => {
         const cliArgs = mocks.mockCliArgs()
         output = cliArgs.output
         const workspace = mocks.mockWorkspace({})
+        workspace.listUnresolvedReferences.mockImplementation(mockedList)
         workspace.getElementIdsBySelectors.mockResolvedValue([new ElemID('salto', 'Account')])
         workspace.flush.mockRejectedValue(new Error('Oy Vey Zmir'))
         result = await cloneAction({
@@ -154,6 +150,7 @@ describe('Element command group', () => {
 
         const cliArgs = mocks.mockCliArgs()
         workspace = mocks.mockWorkspace({})
+        workspace.listUnresolvedReferences.mockImplementation(mockedList)
         workspace.getElementIdsBySelectors.mockResolvedValue([selector])
         result = await cloneAction({
           ...mocks.mockCliCommandArgs(cloneName, cliArgs),
@@ -185,6 +182,7 @@ describe('Element command group', () => {
         const cliArgs = mocks.mockCliArgs()
         output = cliArgs.output
         workspace = mocks.mockWorkspace({})
+        workspace.listUnresolvedReferences.mockImplementation(mockedList)
         workspace.getElementIdsBySelectors.mockResolvedValue([selector])
         result = await cloneAction({
           ...mocks.mockCliCommandArgs(cloneName, cliArgs),
@@ -373,6 +371,7 @@ Cloning the specified elements to inactive.
         const cliArgs = mocks.mockCliArgs()
         output = cliArgs.output
         workspace = mocks.mockWorkspace({})
+        workspace.listUnresolvedReferences.mockImplementation(mockedList)
         result = await moveToEnvsAction({
           ...mocks.mockCliCommandArgs(moveToEnvsName, cliArgs),
           input: {
@@ -407,6 +406,7 @@ Cloning the specified elements to inactive.
         jest.spyOn(callbacks, 'getUserBooleanInput').mockImplementationOnce(() => Promise.resolve(false))
         const cliArgs = mocks.mockCliArgs()
         workspace = mocks.mockWorkspace({})
+        workspace.listUnresolvedReferences.mockImplementation(mockedList)
         workspace.getElementIdsBySelectors = jest.fn().mockResolvedValue([selector])
         result = await moveToEnvsAction({
           ...mocks.mockCliCommandArgs(moveToEnvsName, cliArgs),
@@ -436,6 +436,7 @@ Cloning the specified elements to inactive.
         const cliArgs = mocks.mockCliArgs()
         output = cliArgs.output
         workspace = mocks.mockWorkspace({})
+        workspace.listUnresolvedReferences.mockImplementation(mockedList)
         workspace.getElementIdsBySelectors = jest.fn().mockResolvedValue([selector])
         result = await moveToEnvsAction({
           ...mocks.mockCliCommandArgs(moveToEnvsName, cliArgs),
@@ -478,6 +479,7 @@ Moving the specified elements to envs.
         const cliArgs = mocks.mockCliArgs()
         output = cliArgs.output
         const workspace = mocks.mockWorkspace({})
+        workspace.listUnresolvedReferences.mockImplementation(mockedList)
         workspace.getElementIdsBySelectors.mockResolvedValue([new ElemID('salto', 'Account')])
         workspace.flush.mockRejectedValue(new Error('Oy Vey Zmir'))
         result = await moveToCommonAction({
@@ -508,6 +510,7 @@ Moving the specified elements to envs.
         const cliArgs = mocks.mockCliArgs()
         output = cliArgs.output
         workspace = mocks.mockWorkspace({})
+        workspace.listUnresolvedReferences.mockImplementation(mockedList)
         result = await moveToCommonAction({
           ...mocks.mockCliCommandArgs(moveToCommonName, cliArgs),
           input: {
@@ -543,6 +546,7 @@ Moving the specified elements to envs.
         const cliArgs = mocks.mockCliArgs()
         output = cliArgs.output
         workspace = mocks.mockWorkspace({})
+        workspace.listUnresolvedReferences.mockImplementation(mockedList)
         workspace.getElementIdsBySelectors.mockResolvedValue([selector])
         result = await moveToCommonAction({
           ...mocks.mockCliCommandArgs(moveToCommonName, cliArgs),
@@ -579,6 +583,7 @@ Moving the specified elements to envs.
         const cliArgs = mocks.mockCliArgs()
         output = cliArgs.output
         workspace = mocks.mockWorkspace({})
+        workspace.listUnresolvedReferences.mockImplementation(mockedList)
         workspace.getElementIdsBySelectors.mockResolvedValue([selector])
         result = await moveToCommonAction({
           ...mocks.mockCliCommandArgs(moveToCommonName, cliArgs),
@@ -615,8 +620,6 @@ Moving the specified elements to common.
 
   describe('list-unresolved command', () => {
     const listUnresolvedName = 'list-unresolved'
-    const mockListUnresolved = core.listUnresolvedReferences as jest.MockedFunction<
-      typeof core.listUnresolvedReferences>
 
     describe('success - all unresolved references are found in complete-from', () => {
       let result: CliExitCode
@@ -629,6 +632,7 @@ Moving the specified elements to common.
         userBooleanInput = jest.spyOn(callbacks, 'getUserBooleanInput')
         userBooleanInput.mockRestore()
         workspace = mocks.mockWorkspace({})
+        workspace.listUnresolvedReferences.mockImplementation(mockedList)
         // Should ignore unresolved reference errors
         workspace.errors.mockResolvedValue(mocks.mockErrors([
           new errors.UnresolvedReferenceValidationError({
@@ -659,7 +663,7 @@ Moving the specified elements to common.
       })
 
       it('should call listUnresolvedReferences', () => {
-        expect(core.listUnresolvedReferences).toHaveBeenCalledWith(workspace, 'inactive')
+        expect(workspace.listUnresolvedReferences).toHaveBeenCalledWith('inactive')
       })
 
       it('should print found to console', () => {
@@ -677,7 +681,7 @@ Moving the specified elements to common.
         const cliArgs = mocks.mockCliArgs()
         output = cliArgs.output
         workspace = mocks.mockWorkspace({})
-        mockListUnresolved.mockImplementationOnce(() => Promise.resolve({
+        workspace.listUnresolvedReferences.mockImplementationOnce(() => Promise.resolve({
           found: [],
           missing: [],
         }))
@@ -693,7 +697,7 @@ Moving the specified elements to common.
         expect(result).toBe(CliExitCode.Success)
       })
       it('should call listUnresolvedReferences', () => {
-        expect(core.listUnresolvedReferences).toHaveBeenCalledWith(workspace, undefined)
+        expect(workspace.listUnresolvedReferences).toHaveBeenCalledWith(undefined)
       })
 
       it('should print list to console', () => {
@@ -709,7 +713,7 @@ Moving the specified elements to common.
         const cliArgs = mocks.mockCliArgs()
         output = cliArgs.output
         workspace = mocks.mockWorkspace({})
-        mockListUnresolved.mockImplementationOnce(() => Promise.resolve({
+        workspace.listUnresolvedReferences.mockImplementationOnce(() => Promise.resolve({
           found: [new ElemID('salesforce', 'aaa'), new ElemID('salesforce', 'bbb', 'instance', 'ccc')],
           missing: [new ElemID('salesforce', 'fail')],
         }))
@@ -726,7 +730,7 @@ Moving the specified elements to common.
         expect(result).toBe(CliExitCode.Success)
       })
       it('should call listUnresolvedReferences', () => {
-        expect(core.listUnresolvedReferences).toHaveBeenCalledWith(workspace, 'inactive')
+        expect(workspace.listUnresolvedReferences).toHaveBeenCalledWith('inactive')
       })
 
       it('should print list to console', () => {
@@ -743,7 +747,7 @@ Moving the specified elements to common.
         const cliArgs = mocks.mockCliArgs()
         output = cliArgs.output
         workspace = mocks.mockWorkspace({})
-        mockListUnresolved.mockImplementationOnce(() => {
+        workspace.listUnresolvedReferences.mockImplementationOnce(() => {
           throw new Error('oh no')
         })
 
@@ -760,7 +764,7 @@ Moving the specified elements to common.
         expect(result).toBe(CliExitCode.AppError)
       })
       it('should call listUnresolvedReferences', () => {
-        expect(core.listUnresolvedReferences).toHaveBeenCalledWith(workspace, 'inactive')
+        expect(workspace.listUnresolvedReferences).toHaveBeenCalledWith('inactive')
       })
 
       it('should print the error', () => {
