+++ packages/workspace/test/workspace/workspace.test.ts
@@ -18,7 +18,7 @@ import wu from 'wu'
 import {
   Element, ObjectType, ElemID, Field, DetailedChange, BuiltinTypes, InstanceElement, ListType,
   Values, CORE_ANNOTATIONS, isListType, isInstanceElement, isType, isField,
-  isObjectType, ContainerType, Change, AdditionChange, getChangeElement,
+  isObjectType, ContainerType, Change, AdditionChange, getChangeElement, ReferenceExpression,
 } from '@salto-io/adapter-api'
 import {
   findElement, applyDetailedChanges,
@@ -35,7 +35,7 @@ import { createMockNaclFileSource } from '../common/nacl_file_source'
 import { mockStaticFilesSource } from '../utils'
 import { DirectoryStore } from '../../src/workspace/dir_store'
 import { Workspace, initWorkspace, loadWorkspace, EnvironmentSource,
-  COMMON_ENV_PREFIX } from '../../src/workspace/workspace'
+  COMMON_ENV_PREFIX, UnresolvedElemIDs } from '../../src/workspace/workspace'
 import { DeleteCurrentEnvError,
   UnknownEnvError, EnvDuplicationError, ServiceDuplicationError } from '../../src/workspace/errors'
 
@@ -1955,5 +1955,216 @@ describe('getElementNaclFiles', () => {
     const res = await workspace.getElementNaclFiles(id)
     expect(res).toContain('firstFile.nacl')
     expect(res).not.toContain('secondFile.nacl')
+  })
+})
+
+describe('listUnresolvedReferences', () => {
+  let workspace: Workspace
+  let res: UnresolvedElemIDs
+
+  const createEnvElements = (): Element[] => {
+    const type1 = new ObjectType({
+      elemID: new ElemID('salesforce', 'someType'),
+      fields: {
+        f1: { type: BuiltinTypes.STRING },
+        f2: { type: BuiltinTypes.STRING },
+        f3: { type: BuiltinTypes.STRING },
+      },
+    })
+    const type2 = new ObjectType({
+      elemID: new ElemID('salesforce', 'anotherType'),
+      annotations: { _parent: new ReferenceExpression(type1.elemID) },
+      fields: {
+        f1: { type: type1 },
+        f2: { type: BuiltinTypes.STRING },
+        f3: { type: type1 },
+      },
+    })
+    const inst1 = new InstanceElement(
+      'inst1',
+      type1,
+      {
+        f1: 'aaa',
+        f2: new ReferenceExpression(new ElemID('salesforce', 'someType', 'field', 'f3')),
+        f3: 'ccc',
+      },
+    )
+    const inst2 = new InstanceElement(
+      'inst2',
+      type2,
+      {
+        f1: {
+          f1: 'aaa',
+          f2: 'bbb',
+          f3: new ReferenceExpression(new ElemID('salesforce', 'someType', 'instance', 'inst1', 'f1')),
+        },
+        f3: new ReferenceExpression(new ElemID('salesforce', 'someType', 'instance', 'inst1')),
+      },
+    )
+    return [type1, type2, inst1, inst2]
+  }
+
+  describe('workspace with no references', () => {
+    beforeAll(async () => {
+      const elements = createEnvElements().slice(0, 1)
+      workspace = await createWorkspace(
+        undefined, undefined, undefined, undefined, undefined,
+        {
+          '': {
+            naclFiles: naclFilesSource(mockDirStore(), mockParseCache(), mockStaticFilesSource()),
+          },
+          default: {
+            naclFiles: createMockNaclFileSource(elements),
+            state: createState(elements),
+          },
+          other: {
+            naclFiles: createMockNaclFileSource(elements),
+            state: createState(elements),
+          },
+        }
+      )
+      res = await workspace.listUnresolvedReferences()
+    })
+
+    it('should not find any unresolved references', async () => {
+      expect(res.found).toHaveLength(0)
+      expect(res.missing).toHaveLength(0)
+    })
+  })
+
+  describe('workspace with resolved references', () => {
+    beforeAll(async () => {
+      jest.resetAllMocks()
+      const elements = createEnvElements()
+      workspace = await createWorkspace(
+        undefined, undefined, undefined, undefined, undefined,
+        {
+          '': {
+            naclFiles: naclFilesSource(mockDirStore(), mockParseCache(), mockStaticFilesSource()),
+          },
+          default: {
+            naclFiles: createMockNaclFileSource(elements),
+            state: createState(elements),
+          },
+          other: {
+            naclFiles: createMockNaclFileSource(elements),
+            state: createState(elements),
+          },
+        }
+      )
+      res = await workspace.listUnresolvedReferences('other')
+    })
+
+    it('should not find any unresolved references', () => {
+      expect(res.found).toHaveLength(0)
+      expect(res.missing).toHaveLength(0)
+    })
+  })
+
+  describe('workspace with unresolved references and no complete-from env', () => {
+    beforeAll(async () => {
+      jest.resetAllMocks()
+      const defaultElements = createEnvElements().slice(3)
+      const otherElements = createEnvElements()
+      workspace = await createWorkspace(
+        undefined, undefined, undefined, undefined, undefined,
+        {
+          '': {
+            naclFiles: naclFilesSource(mockDirStore(), mockParseCache(), mockStaticFilesSource()),
+          },
+          default: {
+            naclFiles: createMockNaclFileSource(defaultElements),
+            state: createState(defaultElements),
+          },
+          other: {
+            naclFiles: createMockNaclFileSource(otherElements),
+            state: createState(otherElements),
+          },
+        }
+      )
+      res = await workspace.listUnresolvedReferences()
+    })
+
+    it('should not resolve any references', () => {
+      expect(res.found).toHaveLength(0)
+      expect(res.missing).toEqual([
+        new ElemID('salesforce', 'someType', 'instance', 'inst1'),
+      ])
+    })
+  })
+
+  describe('workspace with unresolved references that exist in other env', () => {
+    beforeAll(async () => {
+      jest.resetAllMocks()
+      const defaultElements = createEnvElements().slice(3)
+      const otherElements = createEnvElements()
+      workspace = await createWorkspace(
+        undefined, undefined, undefined, undefined, undefined,
+        {
+          '': {
+            naclFiles: naclFilesSource(mockDirStore(), mockParseCache(), mockStaticFilesSource()),
+          },
+          default: {
+            naclFiles: createMockNaclFileSource(defaultElements),
+            state: createState(defaultElements),
+          },
+          other: {
+            naclFiles: createMockNaclFileSource(otherElements),
+            state: createState(otherElements),
+          },
+        }
+      )
+      res = await workspace.listUnresolvedReferences('other')
+    })
+
+    it('should successfully resolve all references', () => {
+      expect(res.found).toEqual([
+        new ElemID('salesforce', 'someType', 'field', 'f3'),
+        new ElemID('salesforce', 'someType', 'instance', 'inst1'),
+      ])
+      expect(res.missing).toHaveLength(0)
+    })
+  })
+
+  describe('workspace with unresolved references that do not exist in other env', () => {
+    beforeAll(async () => {
+      jest.resetAllMocks()
+      const defaultElements = createEnvElements().slice(3) as InstanceElement[]
+      defaultElements[0].value = {
+        ...(defaultElements[0] as InstanceElement).value,
+        f3: new ReferenceExpression(new ElemID('salesforce', 'unresolved')),
+      }
+      const otherElements = createEnvElements().slice(1)
+      workspace = await createWorkspace(
+        undefined, undefined, undefined, undefined, undefined,
+        {
+          '': {
+            naclFiles: naclFilesSource(
+              mockDirStore(),
+              mockParseCache(),
+              mockStaticFilesSource(),
+            ),
+          },
+          default: {
+            naclFiles: createMockNaclFileSource(defaultElements),
+            state: createState(defaultElements),
+          },
+          other: {
+            naclFiles: createMockNaclFileSource(otherElements),
+            state: createState(otherElements),
+          },
+        }
+      )
+      res = await workspace.listUnresolvedReferences('other')
+    })
+
+    it('should resolve some of the references', () => {
+      expect(res.found).toEqual([
+        new ElemID('salesforce', 'someType', 'instance', 'inst1', 'f1'),
+      ])
+      expect(res.missing).toEqual([
+        new ElemID('salesforce', 'unresolved'),
+      ])
+    })
   })
 })
